# Performance: add io_binding support for bert benchmark util