--- a/onnxruntime/python/tools/transformers/bert_perf_test.py
+++ b/onnxruntime/python/tools/transformers/bert_perf_test.py
@@ -153,6 +153,7 @@ def onnxruntime_inference_with_io_binding(session, all_inputs, output_names, tes
     device = 'cuda' if test_setting.use_gpu else 'cpu'
     for test_case_id, inputs in enumerate(all_inputs):
         result = session.run(output_names, inputs)
+        results.append(result)
         outputs = {}
         for i in range(len(output_names)):
             outputs[output_names[i]] = result[i]
@@ -160,14 +161,13 @@ def onnxruntime_inference_with_io_binding(session, all_inputs, output_names, tes
         input_tensors, output_tensors = create_input_output_tensors(inputs, outputs, device)
         io_binding = create_io_binding(session, input_tensors, output_tensors)
 
-        # warm up
+        # warm up once
         session.run_with_iobinding(io_binding)
 
-        for i in range(test_setting.test_times):
-            start_time = timeit.default_timer()
-            session.run_with_iobinding(io_binding)
-            latency = timeit.default_timer() - start_time
-            latency_list.append(latency)
+        start_time = timeit.default_timer()
+        session.run_with_iobinding(io_binding)
+        latency = timeit.default_timer() - start_time
+        latency_list.append(latency)
 
     return results, latency_list
 
@@ -210,7 +210,9 @@ def run_one_test(model_setting, test_setting, perf_results, all_inputs, intra_op
 
     all_latency_list = []
     if test_setting.use_io_binding:
-        results, all_latency_list = onnxruntime_inference_with_io_binding(session, all_inputs, output_names, test_setting)
+        for i in range(test_setting.test_times):
+            results, latency_list = onnxruntime_inference_with_io_binding(session, all_inputs, output_names, test_setting)
+            all_latency_list.extend(latency_list)
     else:
         for i in range(test_setting.test_times):
             results, latency_list = onnxruntime_inference(session, all_inputs, output_names)