**Description**: Add io_binding support for bert_perf_test.py

**Motivation and Context**
In the recent benchmark for 1p model, we found there's perf gap between python util (bert_bert_test.py) and c++ based util (onnxruntime_perf_test). ~ (10 ms vs. 5.5 ms)

Sample cmd:
ORT_TENSORRT_INT8_ENABLE=0 ORT_TENSORRT_ENGINE_CACHE_ENABLE=1 ORT_TENSORRT_MAX_WORKSPACE_SIZE=4294967296 ORT_TENSORRT_FP16_ENABLE=1 python bert_perf_test.py --model xyz.onnx --batch_size 32 --sequence_length 128 --intra_op_num_threads 24 --use_gpu --provider tensorrt --opt_level=0 --input_ids_name input_ids --input_mask_name attention_mask

./onnxruntime_perf_test -i "trt_engine_cache_enable|true trt_fp16_enable|true" -t 50 -e tensorrt xyz.onnx -s ./res.txt

After experimenting io_binding for the python util, we are able to bring the number down to 
4.02 ms (not output copy to cpu)
7.22 ms (plus output copy to cpu)

Thus we 'd like to add the --use_io_binding support to bring the perf number comparable to c++ based benchmark util.